[Refactoring] : Fixed Debug compilation warnings.

+++ snake_ml/system/Application.cpp
@@ -90,10 +90,10 @@ void Application::OnInput(InputKey inputKey)
 	}
 }
 
-void Application::Update(double dt)
+void Application::Update(float dt)
 {
 	static uint64_t frameCounter = 0;
-	static double elapsedSeconds = 0.0;
+	static float elapsedSeconds = 0.0;
 	static std::chrono::high_resolution_clock clock;
 	static auto t0 = clock.now();
 
@@ -101,7 +101,7 @@ void Application::Update(double dt)
 	auto t1 = clock.now();
 	auto deltaTime = t1 - t0;
 	t0 = t1;
-	elapsedSeconds += deltaTime.count() * 1e-9;
+	elapsedSeconds += deltaTime.count() * 1e-9f;
 	if (elapsedSeconds > 1.0)
 	{
 		wchar_t buffer[500];
